gen: clarify distinction between directory item and discovery doc
A directory item is the small bit of JSON that you get when you read
Google API discovery URL, with basic information and a link to the
full discovery doc.

Rewrite gen to clarify the distinction between items and discovery
docs.  Previously the same fields of the API struct were populated by
unmarshalling both the item and the discovery doc, which was clever
but hard to follow.

In doing this, it became clear that there was nothing to be gained by
having a separate Item type, so we unmarshal items directly into the
API type as before. But we don't overwrite those fields when we
unmarshal the discovery doc. Also, we only retain fields in the API
type that are needed prior to code generation.

Lastly, we use the discovery doc in the places where it can be
(general API info and auth).

Change-Id: I98f8e3d0e8ef85e51d93687b1fb8e0da9844fca1
Reviewed-on: https://code-review.googlesource.com/9710
Reviewed-by: Ross Light <light@google.com>