gen: distinguish directory items from discovery docs
Add the concept of a directory item to disco. This is the small bit of JSON
that you get when you read Google API discovery URL, with basic information
and a link to the full discovery doc.

Rewrite gen to keep items and discovery docs separate. Previously
the same fields of the API struct were populated by two different
JSON objects, which was clever but hard to follow.

Also, use the discovery doc in the places where it can be (general API
info and auth).

Change-Id: I98f8e3d0e8ef85e51d93687b1fb8e0da9844fca1