gen: add schemas to disco
Support JSON schemas in disco.Document.

Propagate throughout the generator.

For now, we retain the name Schema for a type-with-name. So the code
talks about both Schemas and disco.Schemas, and the latter is often
referred to as a "type". A mess! But I will fix in a later CL that
focuses on renaming.

Also add Features to disco.

Change-Id: Iee096524cbe137109b3548afc065fc46dc1ba54b