gen: unmarshal support for special float values
When a Google API needs to send an infinity or NaN as a float in a JSON payload,
it uses one of the strings "Infinity", "-Infinity" or "NaN".

The generated unmarshaling code used the standard encoding/json unmarshaler, so
it failed on these strings.

We could define a type whose underlying type is float, but with a custom
UnmarshalJSON function that handles the strings. (In fact, we have done so --
internal.JSONFloat64). But if we use this type instead of float64 in
generated code, we would break many clients.

Instead, we keep the generated struct as is, but we wrap it for unmarshaling.
If the type is

  type S struct {
    I int	`json:"i"`
    F float64	`json:"f"`
  }

then we write an UnmarshalJSON for S that defines

  var s1 struct {
    F internal.JSONFloat64 `json:"f"`
    *S
  }

and unmarshals into that. The top-level field F hides S.F, so
the JSON key "f" is unmarshaled into it correctly. Lastly, we
copy s1.F into s.F.

This CL handles float64s and *float64s that occur directly in a struct.
It doesn't handle map[string]float64 and related types. But there are
very few of these. We can do them in a later CL.

This CL does not regenerate the clients.

Fixes #139.

Change-Id: I312dcdb88e290ebbd1bf934db845398917866045
Reviewed-on: https://code-review.googlesource.com/10091
Reviewed-by: Ross Light <light@google.com>