gen: generate Pages method when page token is in request
Generate a Pages method for any RPC that supports pagination.

Previously, a method was recognized as paging if it had a pageToken
query parameter, and a pageToken or nextPageToken response field.

This CL broadens the definition of pagination, as well as the
implementation of the Pages method, to include calls with
a "pageToken" or "nextPageToken" field in the request schema.

Fixes #176.

Change-Id: Ie85cd61b367331536a9baf43a35a800c5139feed