transport: add a way to dial gRPC insecurely.
Previously there was no way to do this and use ClientOptions, since
DialGRPC insists on adding creds.

There is no way for DialGRPC to test if grpc.WithInsecure is one of the
opts, because grpc opts are functions and you can't compare functions.

Another way to accomplish this would be to add WithInsecure to the
google.golang.org/api/option package. This way is more hidden, since
users are less likely to explore the godoc for this package.

Change-Id: Iaf1fd2865416a1606211eeaf4b880d2771d693cd