bundler: allow overcommit to avoid starvation
Previously, AddWait took into account the current item's size when
deciding to block. That meant that (a) large items could starve, and
(b) we needed to repeatedly close and recreate a channel to simulate a
broadcast, in order to ensure that at least one waiter made progress.

Now AddWait ignores the current item's size, blocking only if there is
zero space. This means that the bundler will accept one item over its
memory limit. It also reduces the possibility of starvation to unlucky
scheduling, and avoids the need to re-create a channel.

Change-Id: If457a7316b6938fbe2c172863a39132e8e246392