bundler: use weighted semaphore to enforced byte limit
The weighted semaphore class, since it's context-aware, can replace
the previous complex mechanism involving channels.

It is FIFO, so it also solves the starvation problem.

Using it also brings out a flaw in the original, which is that Add
callers could bypass the queue of AddWait callers. So we disallow
mixing calls. This is fine: any application should want to either get
an error or wait on available space, but not both.

Change-Id: If758576c6284abd8c9638db10600d9a9e6699c07