integration-tests/storage: exit test on RPC error
Change t.Error to t.Fatal when checking that an error is non-nil.

Avoids panics like the one in
https://travis-ci.org/google/google-api-go-client/jobs/240247028.

Though it does not address the root cause of that failure, which would
require retry (and should not be addressed in this client, since it is
obsoleted by cloud.google.com/go/storage).

Change-Id: Ia759bb1e04f8e6f9168b8bd5201b0eea2f05e0e1
Reviewed-on: https://code-review.googlesource.com/13611
Reviewed-by: Ross Light <light@google.com>