generator: do single-chunk upload with a MediaBuffer
To support retry when uploading a single chunk of data,
we use a MediaBuffer to hold the chunk. The MediaBuffer keeps
the chunk in memory, so it can be sent repeatedly.

Previously, gensupport.PrepareUpload returned an io.Reader in this
case.  When the RPC was retried, the Reader had already been read to
completion, so the result was a successful RPC with no data.

As before, when the chunk size is zero, we disable chunking
and use the media io.Reader directly in the call. That means
retry is impossible. It is currently up to the caller to check
that case, since retry isn't handled by this logic.

In followup CLs, I intend to move some of the logic in the generator
into gensupport, and attempt to detect retry in the chunk-size=0 case
and fail.

Fixes GoogleCloudPlatform/google-cloud-go#746.

Change-Id: I9b9eee7a18546815f7fc7715ad01270d1311ca13
Reviewed-on: https://code-review.googlesource.com/16570
Reviewed-by: Ross Light <light@google.com>