option: add WithoutAuthentication
Add an option that allows connecting to a service without authentication.

This is useful for testing, but also for accessing public resources. The
immediate motivation is the ability to use cloud.google.com/go/storage to
access public buckets. (See Google-internal bug 37869545.)

Also, check that WithoutAuthentication is not supplied along with other
options that provide credentials. And while we're there, add a couple of
other checks for incompatible options.

Change-Id: I7bd8f88676e73de422d919309bbc598df4f48100
Reviewed-on: https://code-review.googlesource.com/18470
Reviewed-by: Ross Light <light@google.com>
Reviewed-by: Chris Broadfoot <cbro@google.com>