gen: handle 204 (NoContent)
Rarely, an API returns a 204 status code along with an empty body.
The json decoder returns EOF on the body, so the RPC fails.

Detect the 204 and return with an unfilled response. That response may
be invalid from the API's point of view, but I don't see how the
generated code can do any better. The status code is in the response,
so users can detect the 204 and act on it.

Fixes #195.

Change-Id: Iaf5c7e275cdb429d019890577a97eda4e85bf0a7
Reviewed-on: https://code-review.googlesource.com/19490
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>