gensupport: set GetBody for media upload
Since Go 1.8, you can set the GetBody field of http.Request to
re-create the request body for retryable errors.

This happens automatically for resumable upload, because each chunk
reader is a bytes.Reader, and net/http creates a GetBody for that.
But for single-chunk uploads, we must create the GetBody function
ourselves.

This CL adds the functionality to gensupport in a backwards-compatible
way, and updates the generator. It also regenerate the clients.

Other minor changes in this CL are from the go 1.10 vet checks run
during go test.

Fixes GoogleCloudPlatform/google-cloud-go#811.

Change-Id: I9bcbf71a05850c994e5fd011c0a48e150a889d07
Reviewed-on: https://code-review.googlesource.com/22290
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>