transport/http: add NewTransport method
Provide a way to translate ClientOptions to an http.RoundTripper,
and to insert a base transport other than the default.

This allows you to replace the http.RoundTripper at the
bottom of the "stack" that is created for HTTP connections.

Its only use (so far) is to send HTTP traffic to a proxy. In net/http,
that is done by creating an http.Transport with its Proxy field set,
instead of using http.DefaultTransport. See
cloud.google.com/go/httpreplay.Recorder.Client for an example.

Change-Id: I92ae0fbacc75ded411112b357ec2a1db0a3e4f98
Reviewed-on: https://code-review.googlesource.com/28311
Reviewed-by: Ross Light <light@google.com>