generator: avoid resource field/method name conflicts
It's possible that a resource has both a sub-resource and a method
with the same name.  Since the sub-resource name becomes a field name,
the resulting type has both a field and a method with the same name,
which is not valid Go.

Add an underscore to the sub-resource field name to avoid this.

Change-Id: I040fce7314a960313760f75f8d85058f91082652
Reviewed-on: https://code-review.googlesource.com/30510
Reviewed-by: Jean de Klerk <deklerk@google.com>