grpc: explicitly prevent Retry until we're ready for it
gRPC recently added RPC retry capability [1]. This causes failed
RPCs to be retried. It is configured by the server, but can be
disabled by the client. We should probably adopt this feature in
the future and remove our artisanal retries, but until we do so,
we explicitly disable this feature to prevent client libraries
retrying RPCs that gRPC already retried (m*n situation).

1: https://github.com/grpc/proposal/blob/master/A6-client-retries.md

Change-Id: Ic2d4e7de1966742fb746bb39cd17570bd07653e1
Reviewed-on: https://code-review.googlesource.com/c/34250
Reviewed-by: Jonathan Amsterdam <jba@google.com>