Revert "grpc: explicitly prevent Retry until we're ready for it"
This reverts commit 625cd1887957946515db468ce519bb71fa31fc7f.

Reason for revert: this CL is fine, but we're currently testing some grpc-go problems that involve the user using old versions of grpc-go. The option added in this CL is not supported in older versions of grpc-go, so it becomes a nuisance for users that are trying to pin to old grpc-go versions.

Change-Id: I69118e9aef12c99822cbddb8aa4546e3237dada1
Reviewed-on: https://code-review.googlesource.com/c/34290
Reviewed-by: David Symonds <dsymonds@golang.org>