internal: unflake TestConnectionPool
TestConnectionPool has a race between two goroutines firing on channels. This
refactors that portion of the test to only have one channel fired upon at
a time, and the assertions to happen after a single signal.

This test used to fail 1/5 times: it now is able to run 1000 times without
failure.

Change-Id: I3c6b194eb442fe2040c6a5e1c9f7af13fd0385bb
Reviewed-on: https://code-review.googlesource.com/c/35550
Reviewed-by: Eno Compton <enocom@google.com>