all: remove unsupported experimental libraries
There are a number of libraries that are no longer listed in our discovery
documents [1]. This indicates that the backend team no longer intends on
supporting the clients generated from it. In some cases it's also indicative
of an API being turned down.

We can't remove the packages marked stable (v1, v2, etc) but we can remove
those marked unstable (v1beta1, v2alpha, etc). This CL does so.

Relevant issue: #310

1: https://www.googleapis.com/discovery/v1/apis

Change-Id: If6edc6b03afebdef28e503b04dd2c2a73621e3bb
Reviewed-on: https://code-review.googlesource.com/c/35751
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>