internal: add copyright notices
Fixes failing tests (vet.sh is checking for copyrights).

Updates the generator to add copyrights at the top of packages, and manually
adds the copyrights to packages that are no longer being generated. The
generator always updates packages to include a copyright with the current
year.

In 2019 (and each subsequent year) this will cause generator tests to fail due
to the stubs expecting 2018, and receiving 2019. We should at that point
follow up this commit with a change to makes tests ignore year.

Change-Id: I5799929e5b70161676607a9232650381dc8b6650
Reviewed-on: https://code-review.googlesource.com/c/35750
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Eno Compton <enocom@google.com>
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>