docs: Change Example_webHandler to show variables inline
Currently, users looking at Example_webHandler in godoc do not see the
responsewriter, request, or pagetemplate because only code within in the
function block itself is shown (not the signature, not vars outside).

This CL moves those variables into the code block to attempt to make it
a bit easier to grok at first blush.

Change-Id: I394b369c82253ef6cb0c0bf2f5553377571aeea0
Reviewed-on: https://code-review.googlesource.com/c/38010
Reviewed-by: Tyler Bui-Palsulich <tbp@google.com>
Reviewed-by: kokoro <noreply+kokoro@google.com>