internal: update opencensus, remove go version restriction
Bumps opencensus to a version that does not result in grpc@1.16.0. This
unfortunately does not fix https://github.com/googleapis/google-api-go-client/issues/342
since it requires coordinated tagging (currently scheduling with the OC team).
However, it's a step in the right direction.

Also removes the Go 1.12 compilation target, which should fix
https://github.com/googleapis/google-api-go-client/issues/345.

Fixes #345.

Change-Id: I620c589f22d3310db861d96450b3a945ad24e58a
Reviewed-on: https://code-review.googlesource.com/c/google-api-go-client/+/39231
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Chris Broadfoot <cbro@google.com>