mod: switch to non-alpha future versions of opencensus
The experiment proved mostly successful: we are able to remove
github.com/golang/lint from our mod graph, but we not are able to
GO111MODULE=on go get -u ./... on this repo. Filing an issue in golang/go
for the remaining weirdness, for but now this is still forward progress.

This will be tagged 0.3.1.

Change-Id: I78a890a16175e544bc404d403932b452dbb27646
Reviewed-on: https://code-review.googlesource.com/c/google-api-go-client/+/39711
Reviewed-by: Tyler Bui-Palsulich <tbp@google.com>