gensupport: unflake TestCancelUpload
Without any pause, TestCancelUpload would sometimes blow through all messages
before the test was able to assert and finish. This CL fixes that race by making
each request take 1s, which is more than enough time for the test itself to win
the race over the client/server interactions.

A better solution would be to make it deterministic, but this is a good start.

Fixes #400

Change-Id: If3465a601a71fb73b6a28b8eeaf19ce57a456e50
Reviewed-on: https://code-review.googlesource.com/c/google-api-go-client/+/45671
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Cody Oss <codyoss@google.com>