googleapi: better panic error message if url can't be parsed
Currently, if ResolveRelative is passed something that can't be parsed, the user
sees a segfault. This is not easy to debug. Instead, let's panic with a good
error message.

(unfortunate that the method doesn't return error, but this is the best we can
do)

Also adds a test, and turns the ResolveRelative tests into subtests.

Change-Id: I16005ecc5e200a0bb6712f7d815f462be1d98707
Reviewed-on: https://code-review.googlesource.com/c/google-api-go-client/+/46710
Reviewed-by: Tyler Bui-Palsulich <tbp@google.com>