bundler: limit the total number of goroutines to HandlerLimit
The bundler creates a goroutine per bundle. Only HandlerLimit can
execute in parallel so when write rates are high and handling is
slow, there can be a buildup of goroutines that are competing to
determine if it is their turn to execute. This behavior is
exhibited by TestConcurrentHandlersMax which occasionally fails
to terminate and generates 250,000+ goroutines within 2 seconds
on a 6-core 3.7ghz linux desktop.

In this change, we limit the total number of goroutines to
HandlerLimit. We change the bundle type to be a linked list where
items are added to the tail and handled from the head. Once the
user-specified handler returns within the goroutine, we check
if another bundle is ready for handling. If a bundle is ready,
the existing goroutine handles it. Otherwise, the goroutine ends.
This allows the goroutine count to grow and shrink as necessary.

Since we avoid sending bundles to goroutines prematurely, we
increase our opportunity to fill bundles to their maximum item
count.

This change allows bundler tests to execute and pass when
GOMAXPROCS=1. Previously they would fail to terminate.

Fixes #367
Fixes #417

Change-Id: I73a236ffe49087aedebbbbb6bbc94e0a041e93c1
Reviewed-on: https://code-review.googlesource.com/c/google-api-go-client/+/47991
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jean de Klerk <deklerk@google.com>