Add four simple benchmarks for adding items to a bundle and flushing
The first benchmark measures the cost of Add() only. The second measures the cost
of Add() and Flush() -- waiting for all the added items to be handled. These first
two benchmarks have an immediate handler (no delay).

The third and fourth benchmarks use a handler that takes 1 millisecond (sleeps for
1 millisecond). The third benchmark has bundles of 1 item each, and the fourth
bench is more realistic and has 25 items per bundle.

Also, added one test that checks rough times are right for a handler that has a
300-millisecond delay.

Change-Id: I2d9b8e2971011d1e071f2d05a4d7e2eb5cdd6fc0
Reviewed-on: https://code-review.googlesource.com/c/google-api-go-client/+/47997
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jean de Klerk <deklerk@google.com>