transport: add cert package, default cert source
These certs can be used for mTLS (mutual TLS) connections when Endpoint
Verification is required. A metadata file in a well-known location
signals that a client certificate should be used when opening a
connection to a Google API server.

For more information on Endpoint Validation, see
https://cloud.google.com/endpoint-verification/docs/overview

In the future, the default cert source will be used automatically by
Google API clients. This is an extension of Application Default Credentials.

Change-Id: I66033a84fd735f1ee6dfffdf4f6c0cf4649acc5e
Reviewed-on: https://code-review.googlesource.com/c/google-api-go-client/+/50370
Reviewed-by: Chris Broadfoot <cbro@google.com>