apigee: skip generation
Right now the discovery doc that is generated for this api can't
be turned into generated code. This is blocking all other clients
being updated. The issue is that
organizations.sites.menuitems.save/updatePriorities has a request
body that is a `google.protobuf.Struct`. This does not seem to
map to anything, so the discovery doc does not have a ref schema.
Until the generator for the discovery doc is updated, or the
request body changed, this api will no longer be updated.

Change-Id: I8efa72f664f6bb00fada5b697d90146dd5c5c2a9
Reviewed-on: https://code-review.googlesource.com/c/google-api-go-client/+/54530
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Chris Broadfoot <cbro@google.com>