slides: make int values of Range optional
Depending on `Type` of `Range` speicified the api either expects
or does not expect values. In the case of type `ALL`, these indexes
should not be sent. In the case of type `FIXED_RANGE`, a user might
want to set either the start index or end index to 0. By making
these two pointer types both ways can work provided the correct
input.

Fixes: #433
Change-Id: I060d648dfd72ef51bd8ba202c45dc9a486efed4a
Reviewed-on: https://code-review.googlesource.com/c/google-api-go-client/+/55090
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Chris Broadfoot <cbro@google.com>