customsearch: update custom search and stop generation temporarily
The customsearch changes were generated by running
`./google-api-go-generator -cache=false -install -api=customsearch:v1 -gendir=..`.

Notice, the revision is actually going back in time. This is
expected in this rare case. There were some issues with this apis
spec do to the discovery doc migration. This is being tracked in
b/155117315. For now it seems the old version is the correct one. I
also updated the example to refeclt the old api. This is the reason
this example has broken so much recently.

GH issue #471 was opened to track re-enabling this api.

Change-Id: Id0522dede55fd439aab1438fc701f5d2369a495d
Reviewed-on: https://code-review.googlesource.com/c/google-api-go-client/+/55611
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Tyler Bui-Palsulich <tbp@google.com>