all: only update clients if revision pulled is newer
The discovery doc api can return different revisions of an api spec
with back to back calls. We should not allow our clients to flip
back forth between revisions when we re-generate them. With this
CL we now check to make sure the spec pulled has a newer revison
than the local cache before updating. If the revision is older,
we now skip generating that client.

Fixes: #472
Change-Id: I269ac1a37f74f07bf6d07db6d6213a05227c8f81
Reviewed-on: https://code-review.googlesource.com/c/google-api-go-client/+/55650
Reviewed-by: Tyler Bui-Palsulich <tbp@google.com>
Reviewed-by: kokoro <noreply+kokoro@google.com>