googleapi: return error details
According to aip.dev/193 our api error responses should contain a
details field. We should parse this field and add its contents to
the error string. Without this some of our error messages
are very vague. For example: `googleapi: Error 400: The
request has errors`. This is not actionable for users.

Also, appears there was an idea of having more error details at
one point in time with the Errors field. I don't think this is ever
used in practice though. We might want to consider deprecating this
field in a future release.

Fixes: #473
Change-Id: Iab992a07d3a6a98033ef850f7eb6091815d9e916
Reviewed-on: https://code-review.googlesource.com/c/google-api-go-client/+/55730
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Tyler Bui-Palsulich <tbp@google.com>