compute: backport AllowGlobalAccess field on ForwardingRule
Backporting this field that became present in v0.16.0. This is
being done to help the kubernetes project, which would like to use
this field but can't easily pull in v0.16.0 due to changes made in
the new gRPC version being pulled by that release. Reference issue:
https://github.com/kubernetes/kubernetes/pull/88392.

This commit will be tagged v0.15.1.

Change-Id: I357051506f17173b69f6cc87eae3784d1ad1a7f9