support/bundler: fix deadlock when using time based flush
bundler.Bundler can run into a deadlock under the following
configuration:

- handler limit of 1
- high size and count thresholds such that the only reason a bundle is
flushed is because of the delay timer expiring

Under these conditions, with high concurrency, a deadlock can happen
because the background handler goroutine can exit while there is a
pending queue, which will never get flushed.

Specifically, this sequence of events:

- handler goroutine calls postHandle and receives a nil for the next
bundle, signaling it has no more work to do.

- some other goroutine calls AddWait() -> enqueueCurBundle, which sees
that b.handlerLimit cannot be exceeded and proceeds to queue the item

- handler goroutine decrements b.handlerLimit and exits

The fix is to move the decrement of b.handlerLimit into postHandle such
that the decision to exit and allowing a new handler goroutine to be
spawned cannot be re-ordered.

Fixes: googleapis/google-api-go-client#475

Change-Id: I84421e53a4dc24eec36ceb147933aab57a40aec8
Reviewed-on: https://code-review.googlesource.com/c/google-api-go-client/+/55950
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Cody Oss <codyoss@google.com>