internal: refactor replacement pkgs and remove monitoring v1
Refactored this bit of code so it is a little more flexible. This
now allows more granular control of which versions should be
deprecated.

Monitoring v1 is still recieves updates today and has some apis
that are not captured by v3. Therefore, I am removing the
deprecation notice. This was brought to my attention in
https://github.com/googleapis/google-cloud-go/issues/2213.