internal: fix autosynth
It turns out autosynth creates the PR for you, so we don't have to
do all the extra bits. It should detect that there are local
changes and open a PR.

I am still not sure we will be using this automation in the
future. I would like to simplify things and move over to GH actions
eventually. Because of this I am leaving the code to submit a PR
in for now. I filed #519 to track clean up this code or switching
automation.