gensupport: reintroduce retry deadline
It turns out that ResumableUpload.Upload actually uses
SendRequest to send the request ultimately (via the
transferChunk and doUploadRequest methods). Therefore,
adding a retry to SendRequest effectively removes the
existing deadline for each chunk.

This commit re-adds a deadline like what was in the
original form of #528 as a safety mechanism so that callers
don't see new confusing hangs during the resumable upload.
The test that checked for indefinite retries in an upload
was also re-added.

In the future, it makes sense to either remove the retry
in ResumableUpload.Upload, or revert the changes in SendRequest
and create a new SendAndRetryRequest function instead. But
either of those (especially the first) seems complicated to
do in the immediate term. And the double retry loop we have
here is ultimately harmless because the outer loop will
immediately timeout once the inner loop is done retrying.

Fixes googleapis/google-cloud-go#2469