feat(internal/kokoro): switch off of autosynth onto nightly kokoro
Right now we use autosynth to regenerate nightly. This has been
working okay for us, but does have some downsides.

1. We can't have a programmatic naming of PRs. Right now for each
PR opened we have to manually rename the PR title each time. In
the future regens might auto-merge on green, this would be a
blocker for that.

2. We don't control the build env. Recently, none of our regens
are green from the start. They all need to have `go mod tidy` ran
from an environment of go1.13+. We could add a step in our script
to install go 1.13 everytime we regen, but it seems cleaner to
just run from one of our existing Go docker images with trampoline.

Fixes: #590
Fixes: #519