fix(google-api-go-generator): add patch for compute mtls endpoint (#761)
Compute discovery doc is currently missing MTLSRootURL since compute discovery doc is generated using old apiary pipeline, and it is not easy to fix. This is expected to be fixed after compute migrates to OP discovery doc gen in 2021. Adding a temporary patch for compute in the meantime. (Compute is the only service that has this issue, among the services we care about for MTLS.) Note that I specifically chose not to add a regex-based generic patch here in order to avoid masking MTLS endpoint issues with other services - if there are any issues, we need to triage them on a case by case basis.