fix(idtoken): provide default scope for cert endpoint
When NewValidator is called without any options passed in it will
fail talking to the google cert endpoint because the dailed
authenticated client will not have proper scopes and leads to the
error: "invalid_scope". We should set a default scope do this method
can be called with no extra options.

Fixes: #1187