feat(internal/gensupport): add configurable retry (#1324)
Extend SendAndRetryRequest to allow retry options (backoff and
ErrorFunc) to be passed in. Add a surface to the generated code
for storage to allow the manual layer to choose whether the request
is retried.

I tested this out via the manual layer by verifying that the
ObjectsInsertCall only retries now when WithRetry() is applied.